Party POST Completado

+++ apiv2/models/main.go
@@ -1,8 +1,8 @@
 package models
 
 type Locations struct{
-	Longitude int
-	Latitude int
+	Longitude int64
+	Latitude int64
 }
 
 type Party struct{
